+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * New Module FM16SEQ: fm synth for every step, randomization options
   * New Module FRACT: delay for fractions of a beat
+  * Buffer: added stereo in/out, added chance to freeze
   * added scales to all quantizers - this should not break existing patches
   * Play mode knob tooltip is better now for all modules